requirements.txt
fastapi
sqlalchemy
uvicorn
python-multipart
alembic
jinja2
passlib
bcrypt
python-jose[cryptography]